requirements.txt
streamlit
torch==1.12.1+cpu
torchvision==0.13.1+cpu
numpy>=1.19.2
opencv-python
Pillow>=8.4.0